Found final visible bug. IFDEF on LONG_MAX seems like a bad idea

+++ FDK/Tools/Programs/makeotf/makeotf_lib/source/pstoken/pstoken.c
@@ -645,7 +645,7 @@ finish:
 
 /* Match token */
 int psMatchToken(psCtx h, psToken *token, int type, char *value) {
-	long length = strlen(value);
+	size_t length = strlen(value);
 	return token->type == type && token->length == length &&
 	       memcmp(&h->cb.buf->array[token->index], value, length) == 0;
 }
@@ -665,7 +665,7 @@ psToken *psFindToken(psCtx h, int type, char *value) {
 
 /* Match token's value */
 int psMatchValue(psCtx h, psToken *token, char *value) {
-	long length = strlen(value);
+	size_t length = strlen(value);
 	return token->length == length &&
 	       memcmp(&h->cb.buf->array[token->index], value, length) == 0;
 }
