+++ FDK/Tools/Programs/makeotf/makeotf_lib/source/typecomp/parse.c
@@ -1646,7 +1646,7 @@ static void readFontName(parseCtx h) {
 }
 
 /* Read integer-valued dict key */
-static long readIntKey(parseCtx h, int iKey) {
+static int32_t readIntKey(parseCtx h, int iKey) {
 	psToken *token = psGetToken(h->ps);
 	if (token->type != PS_INTEGER) {
 		badKeyValue(h, iKey);
@@ -1657,7 +1657,7 @@ static long readIntKey(parseCtx h, int iKey) {
 /* Process literal */
 static void doLiteral(parseCtx h, psToken *literal) {
 	tcCtx g = h->g;
-	long type;
+	int32_t type;
 	String key;
 	DictKeyMap *map;
 
@@ -1899,7 +1899,7 @@ static void calcUDV(parseCtx h, double *BDM, double *BDP,
 }
 
 /* Get integer value from dict key */
-static long getKeyInt(parseCtx h, int iKey, int required) {
+static int32_t getKeyInt(parseCtx h, int iKey, int required) {
 	psToken *token = &h->keys[iKey].value;
 
 	if (!SEEN_KEY(iKey)) {
@@ -2004,7 +2004,7 @@ static void saveMM(parseCtx h, DICT *dict, int iKey) {
 			order[i] = 0;
 			for (j = 0; j < nAxes; j++) {
 				double d = BDP[k++];
-				long n = (long)d;
+				int32_t n = (int32_t)d;
 
 				if (n != d || (n != 0 && n != 1)) {
 					/* Non-integer master coordinates must have CDV support */
@@ -2234,7 +2234,7 @@ static void saveNumber(parseCtx h, DICT *dict, int iKey) {
 	psToken *token = &key->value;
 
 	if (token->type == PS_INTEGER) {
-		long i = psConvInteger(h->ps, token);
+		int32_t i = psConvInteger(h->ps, token);
 		if (key->dflt != NULL && i == strtol(key->dflt, NULL, 0)) {
 			return; /* Matched default */
 		}
@@ -2713,7 +2713,7 @@ static int defaultNumber(parseCtx h, int iKey) {
 	psToken *token = &key->value;
 
 	if (token->type == PS_INTEGER) {
-		long i = psConvInteger(h->ps, token);
+		int32_t i = psConvInteger(h->ps, token);
 		if (key->dflt != NULL && i == strtol(key->dflt, NULL, 0)) {
 			return 1; /* Matched default */
 		}
@@ -2748,7 +2748,7 @@ static void saveDicts(parseCtx h) {
 
 	if (!SEEN_KEY(iStdVW)) {
 		/* StdVW not found (old font) so use 16th erode proc token! */
-		long StdVW = getStdVWFromErodeProc(h);
+		int32_t StdVW = getStdVWFromErodeProc(h);
 		if (StdVW != -1) {
 			dictSaveInt(&h->font->Private, StdVW);
 			DICTSAVEOP(h->font->Private, cff_StdVW);
